+++ junifer/preprocess/fsl/__init__.py
@@ -2,5 +2,3 @@
 
 # Authors: Synchon Mandal <s.mandal@fz-juelich.de>
 # License: AGPL
-
-from .bold_warper import BOLDWarper
